backend/lambdas/layers/simulation/pyproject.toml
[tool.poetry]
name = "simulation"
version = "0.1.0"
description = ""
authors = ["Adam Ayre-Storie <adam.ayre-storie@hydrologiq.com>"]
readme = "README.md"
include = ["simulation/simulation/query/queries/hydrogen_nrmm*.*"]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
linkml-runtime = "^1.6.3"
geopy = "^2.4.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
mock = "^5.1.0"
requests-mock = "^1.11.0"
python-dotenv = "^1.0.0"
jmespath = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "api_run_simulation"